user-service/requirements.txt
Flask
Flask-SQLAlchemy
PyJWT
Werkzeug
pymysql
python-dotenv
cryptography
